new pos encoding

+++ src/data/utils.py
@@ -1,3 +1,4 @@
+import numpy as np
 import torch
 from collections import defaultdict
 
@@ -23,6 +24,9 @@ class Collater:
         for inp in inputs:
             for key, obj in inp.items():
 
+                #tmp =  np.log( * 1e2) / 10
+                
+
                 grouped[key]["flat_tensor"].append(
                     torch.as_tensor(obj["flat_tensor"], dtype=torch.float32)
                 )
@@ -33,7 +37,7 @@ class Collater:
                     torch.as_tensor(obj["local_patch_ids"], dtype=torch.long)
                 )
                 grouped[key]["patch_positions"].append(
-                    torch.as_tensor(obj["patch_positions"], dtype=torch.long)
+                    torch.as_tensor(obj["patch_positions"], dtype=torch.float32)
                 )
         
         # ------------------------------------------------------------
@@ -44,10 +48,10 @@ class Collater:
         for key, obj in grouped.items():
            
             out[key] = {
-                "flat_tensor": torch.stack(obj["flat_tensor"], dim=0),  # (B, P, C)
-                "global_patch_ids": torch.stack(obj["global_patch_ids"], dim=0),      # (B, P)
-                "local_patch_ids": torch.stack(obj["local_patch_ids"], dim=0),      # (B, P, 3)
-                "patch_positions": torch.stack(obj["patch_positions"], dim=0),      # (B, P, 3)
+                "flat_tensor": torch.stack(obj["flat_tensor"], dim=0).to(torch.float32),  # (B, P, C)
+                "global_patch_ids": torch.stack(obj["global_patch_ids"], dim=0).to(torch.long),      # (B, P)
+                "local_patch_ids": torch.stack(obj["local_patch_ids"], dim=0).to(torch.long),      # (B, P, 3)
+                "patch_positions": torch.stack(obj["patch_positions"], dim=0).to(torch.float32),      # (B, P, 3)
             }
             
             axis_sum = torch.sum(torch.abs(out[key]["flat_tensor"]), dim=2)
